Add files via upload
The objective of an online tourism website is to implement a targeted marketing strategy, transitioning from a broad outreach to a digital approach, in collaboration with a social networking platform. The aim is to understand and leverage customers' digital and social behavior to deliver targeted digital advertisements to users who exhibit a high likelihood of purchasing travel products.

Given the relatively higher cost of digital platform advertisements, precision in model creation becomes paramount. The emphasis is on achieving high accuracy to ensure that the marketing budget is effectively utilized and the digital advertisements effectively reach users with a strong inclination to make travel-related purchases.

Recognizing that the propensity to buy tickets varies across different login devices, the task involves developing two separate models—one for laptop users and another for mobile users (where anything other than a laptop is considered as mobile phone usage). This distinction is crucial due to the observed differences in purchasing behavior between these device categories.